move utils for building parameters from cards to own utils file
