--- frontend/src/metabase/meta/Parameter.js
@@ -1,41 +1,0 @@
-import MetabaseSettings from "metabase/lib/settings";
-import type { TemplateTag } from "metabase-types/types/Query";
-import type { Parameter } from "metabase-types/types/Parameter";
-
-const areFieldFilterOperatorsEnabled = () =>
-  MetabaseSettings.get("field-filter-operators-enabled?");
-
-// NOTE: this should mirror `template-tag-parameters` in src/metabase/api/embed.clj
-export function getTemplateTagParameters(tags: TemplateTag[]): Parameter[] {
-  function getTemplateTagType(tag) {
-    const { type } = tag;
-    if (type === "date") {
-      return "date/single";
-    } else if (areFieldFilterOperatorsEnabled() && type === "string") {
-      return "string/=";
-    } else if (areFieldFilterOperatorsEnabled() && type === "number") {
-      return "number/=";
-    } else {
-      return "category";
-    }
-  }
-
-  return tags
-    .filter(
-      tag =>
-        tag.type != null && (tag["widget-type"] || tag.type !== "dimension"),
-    )
-    .map(tag => {
-      return {
-        id: tag.id,
-        type: tag["widget-type"] || getTemplateTagType(tag),
-        target:
-          tag.type === "dimension"
-            ? ["dimension", ["template-tag", tag.name]]
-            : ["variable", ["template-tag", tag.name]],
-        name: tag["display-name"],
-        slug: tag.name,
-        default: tag.default,
-      };
-    });
-}
